backend/requirements.txt
fastapi
uvicorn
fastapi_azure_auth
typing
pydantic
pydantic_settings
httpx
mysql-connector-python